# Conflicts:
#	HalAe/.idea/caches/build_file_checksums.ser
#	HalAe/app/src/main/AndroidManifest.xml

+++ HalAe/app/src/main/java/com/team/halae/RecommendDonationData.kt
@@ -6,5 +6,5 @@ data class RecommendDonationData (
         var don_img : String,
         var hal_name : String,
         var don_now : Int,
-        var don_percent : Int
+        var don_percent : Float
 )
